refactor(LottoGenerator): 느슨한 결합을 위한 인터페이스 및 구현체 사용
